+++ HISTORY.md
@@ -1,5 +1,15 @@
 # Revision History
 
+## Version 2.3.0 / xx-Nov-2018
+
+* Change: Better determination of hostname stored in ABRequest
+* Change: Remove method terminate() for consistency
+* Add: Include component version in error JSON object
+* Add: Include debug messages for function invocation, e.g. sendHTML(), getStatistics()
+* Add: Include readFile() method to read content of a file
+* Add: Add the HTTP and HTTPS ports in the statistics object
+* Add: Add the HTTP and HTTPS ports in the toString() method
+
 ## Version 2.2.0 / 19-Nov-2018
 
 * Add: Include sample code HelloWorld.ts, GetPost.ts, FileServer.ts and RedirectHttps.ts